[style] #19 ViewHolder 분리
